+++ R/treeio.R
@@ -4,6 +4,6 @@ getRoot <- tidytree:::rootnode
 get.tree <- treeio::get.tree
 drop.tip <- treeio::drop.tip
 get.fields <- treeio::get.fields
-is.tree <- treeio:::is.tree
+is.tree <- getFromNamespace('is.tree', 'treeio')
 
 set_branch_length <- getFromNamespace("set_branch_length", "treeio")